bubble sort

+++ 0-bubble_sort.c
@@ -6,7 +6,7 @@
  */
 void bubble_sort(int *array, size_t size)
 {
-unsigned int n, j, i;
+size_t n, j, i;
 for (i = 0; i < size - 1; i++)
 {
 for (j = 0; j < size - i - 1; j++)
